fix(app): Fix eslint config and fix file lintings
The eslint config was not linting the files properly.
It turns out that the line: "prettier/prettier": ["error"]
is missing in the rules of the eslint config. Fix this so
the linter works properly.

Now that the linter is linting properly, fix the files that are
violating the linting rules.